Kamila e Kevin 3AI

+++ src/main/java/br/com/etechoracio/ingresso/dto/SessaoResponseDTO.java
@@ -9,7 +9,7 @@ public record SessaoResponseDTO(
         Long id,
         LocalDate data,
         LocalTime horario,
-        Double preco,
+        float preco,
         String sala,
         Filme filme
 ) {}
